Finished basic testing for admin auhtor and borrwer controllers

+++ src/main/java/com/ss/training/lms/controller/AdminBorrwerController.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -20,16 +21,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping(path="lms/admin")
 public class AdminBorrwerController {
 
     @Autowired
     AdminBorrowerService adminBorrowerService;
 
-    @PostMapping(path = "/lms/admin/borrowers/add",
-    produces = {
-        MediaType.APPLICATION_JSON_VALUE,
-        MediaType.APPLICATION_XML_VALUE
-    })
+    @PostMapping(path = "/borrowers")
     public ResponseEntity<Borrower> addABorrower(@RequestBody Borrower borrower) {
         HttpStatus status = HttpStatus.CREATED;
 
@@ -48,11 +46,7 @@ public class AdminBorrwerController {
     }
     
 
-    @PostMapping(path = "/lms/admin/borrowers/delete",
-    produces = {
-        MediaType.APPLICATION_JSON_VALUE,
-        MediaType.APPLICATION_XML_VALUE
-    })
+    @DeleteMapping(path = "/borrowers")
     public ResponseEntity<Borrower> deleteABorrower(@RequestBody Borrower borrower) {
         HttpStatus status = HttpStatus.ACCEPTED;
 
@@ -61,6 +55,12 @@ public class AdminBorrwerController {
             return new ResponseEntity<Borrower>(borrower, status);
         }
         try {
+        	Borrower checkTable = adminBorrowerService.readABorrower(borrower.getCardNo());
+        	if (checkTable == null)
+        	{
+        		status = HttpStatus.NOT_FOUND;
+                return new ResponseEntity<Borrower>(borrower, status);
+        	}
             adminBorrowerService.deleteABorrower(borrower);
         } catch (SQLException e) {
             e.printStackTrace();
@@ -70,11 +70,7 @@ public class AdminBorrwerController {
         return new ResponseEntity<Borrower>(borrower, status);
     }
 
-    @PutMapping(path = "/lms/admin/borrowers/update",
-        produces = {
-            MediaType.APPLICATION_JSON_VALUE,
-            MediaType.APPLICATION_XML_VALUE
-        })
+    @PutMapping(path = "/borrowers")
     public ResponseEntity<Borrower> updateABorrower(@RequestBody Borrower borrower) {
         HttpStatus status = HttpStatus.ACCEPTED;
 
@@ -83,6 +79,12 @@ public class AdminBorrwerController {
             return new ResponseEntity<Borrower>(borrower, status);
         }
         try {
+        	Borrower findBorrower = adminBorrowerService.readABorrower(borrower.getCardNo());
+        	if (findBorrower == null)
+        	{
+        		status = HttpStatus.NOT_FOUND;
+                return new ResponseEntity<Borrower>(borrower, status);
+        	}
             adminBorrowerService.updateABorrower(borrower);
         } catch (SQLException e) {
             e.printStackTrace();
@@ -92,7 +94,7 @@ public class AdminBorrwerController {
         return new ResponseEntity<Borrower>(borrower, status);
     }
 
-    @RequestMapping(path = "/lms/admin/borrowers/{cardNo}")
+    @RequestMapping(path = "/borrowers/{cardNo}")
     public ResponseEntity<Borrower> readABorrower(@PathVariable int cardNo) {
         HttpStatus status = HttpStatus.OK;
         Borrower borrower = null;
@@ -112,7 +114,7 @@ public class AdminBorrwerController {
         return new ResponseEntity<Borrower>(borrower, status);
     }
 
-    @RequestMapping(path = "/lms/admin/borrowers")
+    @RequestMapping(path = "/borrowers")
     public ResponseEntity<List<Borrower>> readAllBorrowers() {
         HttpStatus status = HttpStatus.OK;
         List<Borrower> borrowers = null;
